Fix the boost gauge so it reports boost pressure accurately as well as indicating the boost pressure more responsively on the gauge face.
Related Work Items: #11

+++ src/src/Config/DisplayConfiguration.cs
@@ -90,7 +90,7 @@
         /// <value>
         /// The maximum boost.
         /// </value>
-        public int MaxBoost { get; set; } = 15;
+        public double MaxBoost { get; set; } = 15;
 
         /// <summary>
         /// Gets or sets the maximum idle load.
